Clear XSRF token when navigating to Login
This will fix the WebSocket authentication error if redirected to
the Login page when a session is expired. The WebSocketPlugin checks
to see if the user is logged in before attempting to make a connection.
The AuthenticationStore determines if a user is logged in
based on whether or not the XSRF-TOKEN cookie exists. Currently the
app only removes this token if the user explicitly logs out from
the application header.

Signed-off-by: Yoshie Muranaka <yoshiemuranaka@gmail.com>
Change-Id: I9f2b7025ff1de21889267b8f56efd08930eb63d3