Remove envConstants.js file
Removing because current implementations are not using this file but
directly referencing process.env.VUE_APP_* value instead.

Also removing reference to this file in the env documentation along with
the conditional template rendering section. The conditional template pattern
is something that should be avoided and currently researching ways to
dynamically generate navigation items so we can avoid this pattern.

Signed-off-by: Yoshie Muranaka <yoshiemuranaka@gmail.com>
Change-Id: Icbb92c423dac2b9b353c3701c330c3b9bfb00d7a