Update component documentation
The component structure was not consistent. We determined how we
wanted to save image examples within the component directory, but not
all of the components were updated. The result was that some images
were not displaying on the component pages. This patchset resolves
that issue, along with removing the page component since it is not an
actual component within the application.

- Add directory for each component, move example images to the
directory, and update the image path
- Create a page-title and page-section directory and index.md for each
- Move content for page-title and page-section from the page.md file
into the index.md within the respective component directory
- Delete the page.md from the component directory as it is not a
component
- Update links to page.md in the page-anatomy.md file.

Signed-off-by: Derick Montague <derick.montague@ibm.com>
Change-Id: I1e554adf71abb4c84f423a30d3c3b598f678ade0