Update PageContainer snapshot test
The UI is rendering differently than the snapshot saved. I reviewed
the difference and it is simply a difference with how the formatting
of the class and id values. I did not see this issue when testing the
PageContainer snapshot test. It is not clear why there is a
difference, but if it were to fail again for a different user, we can
remove the snapshot since the main test we want to pass is that the
HTML element being rendered is the <main> element.

To update the snapshot for an individual spec file, use the following
command in the cli:

npm run test:update -- -u tests/unit/Global/PageContainer.spec.js

Signed-off-by: Derick Montague <derick.montague@ibm.com>
Change-Id: Ieac495f893ba2025d2d6f4d3b7b49506bdf0ecae