Add .gitattributes file
This file has a setting that assures that our end of line defaults to
lf for all developers when cloning or checking out a branch. This is
important for windows users. When they checkout, the eol is being set
to CRLF which slows down the build and requires they run the lint
command and stage every file in the application to keep linting from
slowing down the build process.

https://prettier.io/docs/en/options.html#end-of-line

Signed-off-by: Derick Montague <derick.montague@ibm.com>
Change-Id: I700e5f1591437c60868ac7e81bd29d613ad868e9