Update all minor and patch releases
Updated NPM packages with minor and patch releases only.

Tested the UI locally and by building copying to the BMC and running in
an overlay and did not observe breaking changes.

Major releases are not updated in this commit as they require
additional testing and code updates to address breaking changes and will
be updated separately once those updates are stable.


Signed-off-by: Dixsie Wolmers <dixsie@ibm.com>
Change-Id: I73c952a75a1148c5852a18d73b9065083fd8e94c