Change filtering virtual media
When Virtual Media image was configured and connected entry
did not appear on the UI after opening in the new tab or
after page refresh. The cause was filtering predicate.
Proxy devices has transferProtocolType eq to 'OEM',
legacy the onces that has no transferProtocolType. For
mouting from the share the property exists and has value of
'CIFS'. I added object with all protocols (see docs), and
change the filtering for legacy devices. New condition is
all accept 'OEM'.

Signed-off-by: Grzegorz Raguza <grzegorzx.raguza@intel.com>
Change-Id: I7ba123f19f924536f131ec40d15715b7f50fc985