Add ability to enable DHCP (#60)
* Add ability to enable DHCP

Saving to the first interface automatically updates DHCPv4
and DHCPv6 on all interfaces.

Enabling DHCP sends a request to the server to return a DHCP IP
address. If DHCP is enabled and the DHCP network is not configured,
then the system will go down and network settings will need to be restored

TODO - determine alert or confirmation modal text to let user know the
system could go down.

Signed-off-by: Dixsie Wolmers <dixsie@ibm.com>

* undo change of UseNTPServer to UseDNSServers