Add timeout mechanism when setting Server power policy (#68)
When setting the Server power policy through the GUI, because there
is a delay in updating the DBus, it cannot return immediately, and
the GUI will immediately obtain the property value through the Get
method after setting, resulting in the obtained value is not the
updated value, resulting in The radio is not updated.
Add a delay of 2s to solve this problem.

Signed-off-by: George Liu <liuxiwei@inspur.com>

Signed-off-by: George Liu <liuxiwei@inspur.com>